requirements.txt
tenacity
langid
jproperties
sentence-transformers
PyMuPDF
python-docx